backend/requirements.txt
google-generativeai>=0.3.0
groq>=0.4.0
huggingface-hub>=0.20.0
together>=0.2.0
cohere>=4.0.0
litellm>=1.17.0
beautifulsoup4>=4.12.0
requests>=2.31.0
pandas>=2.0.0
numpy>=1.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0
feedparser>=6.0.0
praw>=7.7.0
PyMuPDF>=1.23.0
tenacity>=8.2.0
lxml>=4.9.0
aiohttp>=3.9.0
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.5
